Refactor: directory structure 수정
