+++ todang/src/main/java/com/jichijima/todang/service/user/UserService.java
@@ -1,8 +1,8 @@
-package com.jichijima.todang.service;
+package com.jichijima.todang.service.user;
 
 import com.jichijima.todang.model.dto.user.UserUpdateRequest;
-import com.jichijima.todang.model.entity.User;
-import com.jichijima.todang.repository.UserRepository;
+import com.jichijima.todang.model.entity.user.User;
+import com.jichijima.todang.repository.user.UserRepository;
 import com.jichijima.todang.util.JwtUtil;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
